requirements.txt
# Streamlit
streamlit>=1.28.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
camelot-py[cv]>=0.11.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# AI Providers
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# For Camelot dependencies
opencv-python-headless>=4.8.0
ghostscript>=0.7
